@@ -52,12 +52,12 @@
     "postcss-cssnext": "2.11.0",
     "prop-types": "=15.5.10",
     "qrcode.react": "0.7.2",
-    "react": "=15.6.x",
-    "react-animate-on-change": "^1.0.0",
+    "react": "=16.0.0",
+    "react-animate-on-change": "=1.0.0",
     "react-circular-progressbar": "0.2.1",
     "react-css-themr": "=2.1.2",
-    "react-dom": "=15.6.x",
-    "react-i18next": "^5.2.0",
+    "react-dom": "=16.0.0",
+    "react-i18next": "=6.0.6",
     "react-redux": "5.0.6",
     "react-router": "4.2.0",
     "react-router-dom": "4.2.2",
