@@ -1,5 +1,6 @@
 import actionTypes from '../constants/actions';
 import { getForgedBlocks, getForgedStats } from '../utils/api/forging';
+import { errorAlertDialogDisplayed } from './dialog';
 
 export const forgedBlocksUpdated = data => ({
   data,
@@ -11,7 +12,10 @@ export const fetchAndUpdateForgedBlocks = ({ activePeer, limit, offset, generato
     getForgedBlocks(activePeer, limit, offset, generatorPublicKey)
       .then(response =>
         dispatch(forgedBlocksUpdated(response.blocks)),
-      );
+      )
+      .catch((error) => {
+        dispatch(errorAlertDialogDisplayed({ text: error.message }));
+      });
   };
 
 export const forgingStatsUpdated = data => ({
@@ -24,5 +28,8 @@ export const fetchAndUpdateForgedStats = ({ activePeer, key, startMoment, genera
     getForgedStats(activePeer, startMoment, generatorPublicKey)
       .then(response =>
         dispatch(forgingStatsUpdated({ [key]: response.forged })),
-      );
+      )
+      .catch((error) => {
+        dispatch(errorAlertDialogDisplayed({ text: error.message }));
+      });
   };
