@@ -1,9 +1,10 @@
 import { expect } from 'chai';
 import sinon from 'sinon';
 import actionTypes from '../constants/actions';
-import * as saveAccountUtils from '../utils/saveAccount';
+import * as savedAccountsUtils from '../utils/savedAccounts';
 import {
   accountSaved,
+  accountSwitched,
   accountRemoved,
   accountsRetrieved,
 } from './savedAccounts';
@@ -18,12 +19,19 @@ describe('actions: savedAccount', () => {
 
   describe('accountsRetrieved', () => {
     it('should create an action to retrieved the saved accounts list', () => {
-      sinon.stub(saveAccountUtils, 'getSavedAccount').returns([data]);
+      const getSavedAccountsStub = sinon.stub(savedAccountsUtils, 'getSavedAccounts').returns([data]);
+      const getLastActiveAccountStub = sinon.stub(savedAccountsUtils, 'getLastActiveAccount').returns(data);
       const expectedAction = {
-        data: [data],
+        data: {
+          accounts: [data],
+          lastActive: data,
+        },
         type: actionTypes.accountsRetrieved,
       };
       expect(accountsRetrieved()).to.be.deep.equal(expectedAction);
+
+      getSavedAccountsStub.restore();
+      getLastActiveAccountStub.restore();
     });
   });
 
@@ -34,6 +42,16 @@ describe('actions: savedAccount', () => {
         type: actionTypes.accountSaved,
       };
       expect(accountSaved(data)).to.be.deep.equal(expectedAction);
+    });
+  });
+
+  describe('accountSwitched', () => {
+    it('should create an action to save account', () => {
+      const expectedAction = {
+        data,
+        type: actionTypes.accountSwitched,
+      };
+      expect(accountSwitched(data)).to.be.deep.equal(expectedAction);
     });
   });
 
