@@ -45,11 +45,11 @@ describe('Header', () => {
     expect(wrapper.find(Button)).to.have.length(1);
   });
 
-  it('renders 7 RelativeLink components', () => {
-    expect(wrapper.find(RelativeLink)).to.have.length(7);
+  it('renders 10 RelativeLink components', () => {
+    expect(wrapper.find(RelativeLink)).to.have.length(10);
   });
 
-  it('should have an image with srouce of "logo"', () => {
+  it('should have an image with source of "logo"', () => {
     expect(wrapper.contains(<img className={styles.logo} src={logo} alt="logo" />))
       .to.be.equal(true);
   });
