@@ -4,11 +4,11 @@ import Input from 'react-toolbox/lib/input';
 import Dropdown from 'react-toolbox/lib/dropdown';
 import Button from 'react-toolbox/lib/button';
 import i18next from 'i18next';
-import getNetworks from './networks';
 import PassphraseInput from '../passphraseInput';
 import styles from './login.css';
 import env from '../../constants/env';
 import networks from '../../constants/networks';
+import getNetwork from '../../utils/getNetwork';
 import LanguageDropdown from '../languageDropdown';
 import RelativeLink from '../relativeLink';
 import { validateUrl, getLoginData } from '../../utils/login';
@@ -24,7 +24,7 @@ class Login extends React.Component {
     this.state = {
       passphrase: '',
       address: '',
-      network: networks.mainnet,
+      network: networks.mainnet.code,
     };
 
     this.validators = {
@@ -43,8 +43,8 @@ class Login extends React.Component {
   }
 
   getNetworksList() {
-    this.networks = getNetworks().map((network, index) => ({
-      label: i18next.t(network.name),
+    this.networks = Object.keys(networks).map((network, index) => ({
+      label: i18next.t(networks[network].name),
       value: index,
     }));
   }
@@ -64,8 +64,8 @@ class Login extends React.Component {
   }
 
   onLoginSubmission(passphrase) {
-    const network = Object.assign({}, getNetworks()[this.state.network]);
-    if (this.state.network === networks.customNode) {
+    const network = Object.assign({}, getNetwork(this.state.network));
+    if (this.state.network === networks.customNode.code) {
       network.address = this.state.address;
     }
 
@@ -127,10 +127,10 @@ class Login extends React.Component {
 
   autoLogin() {
     const { savedAccounts } = this.props;
-    if (savedAccounts && savedAccounts.length > 0 && !this.props.account.afterLogout) {
-      this.account = savedAccounts[0];
-      const network = Object.assign({}, getNetworks()[this.account.network]);
-      if (this.account.network === networks.customNode) {
+    if (savedAccounts && savedAccounts.lastActive && !this.props.account.afterLogout) {
+      this.account = savedAccounts.lastActive;
+      const network = Object.assign({}, getNetwork(this.account.network));
+      if (this.account.network === networks.customNode.code) {
         network.address = this.account.address;
       }
 
@@ -161,7 +161,7 @@ class Login extends React.Component {
                 className={`${styles.network} network`}
               />
               {
-                this.state.network === networks.customNode &&
+                this.state.network === networks.customNode.code &&
                   <Input type='text'
                     label={this.props.t('Node address')}
                     name='address'
@@ -186,7 +186,7 @@ class Login extends React.Component {
                   <Button label={this.props.t('Login')} primary raised
                     className='login-button'
                     type='submit'
-                    disabled={(this.state.network === networks.customNode && this.state.addressValidity !== '') ||
+                    disabled={(this.state.network === networks.customNode.code && this.state.addressValidity !== '') ||
                     this.state.passphraseValidity !== ''} />
                 </div>
               </footer>
