@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import styles from './offlineWrapper.css';
 
 export const OfflineWrapperComponent = props => (
-  <span className={props.offline && styles.isOffline}>
+  <span className={!props.offline ? '' : styles.isOffline}>
     { props.children }
   </span>
 );
