@@ -38,12 +38,12 @@ class PassphraseGenerator extends React.Component {
 
   /**
    * Tests useragent with a regexp and defines if the account is mobile device
+   * it is on class so that we can mock it in unit tests
    *
    * @param {String} [agent] - The useragent string, This parameter is used for
    *  unit testing purpose
    * @returns {Boolean} - whether the agent represents a mobile device or not
    */
-  // it is on class so that we can mock it in unit tests
   // eslint-disable-next-line class-methods-use-this
   isTouchDevice(agent) {
     return (/(android|bb\d+|meego).+mobile|avantgo|bada\/|blackberry|blazer|compal|elaine|fennec|hiptop|iemobile|ip(hone|od)|iris|kindle|lge |maemo|midp|mmp|mobile.+firefox|netfront|opera m(ob|in)i|palm( os)?|phone|p(ixi|re)\/|plucker|pocket|psp|series(4|6)0|symbian|treo|up\.(browser|link)|vodafone|wap|windows ce|xda|xiino|android|ipad|playbook|silk/i.test(agent || navigator.userAgent || navigator.vendor || window.opera));
@@ -88,10 +88,11 @@ class PassphraseGenerator extends React.Component {
   }
 
   render() {
+    const isTouch = this.isTouchDevice(this.props.agent);
     return (
       <div className={`${grid.row} ${grid['center-xs']}`} >
         <div className={grid['col-xs-12']}>
-          {this.isTouchDevice() ?
+          {isTouch ?
             <div>
               <p>Enter text below to generate random bytes</p>
               <Input onChange={this.seedGeneratorBoundToThis}
