@@ -49,44 +49,44 @@ describe('Send', () => {
 
   it('accepts valid amount', () => {
     wrapper.find('.amount input').simulate('change', { target: { value: '120.25' } });
-    expect(wrapper.find('.amount').text()).to.not.contain('Invalid');
+    expect(wrapper.find('Input.amount').text()).to.not.contain('Invalid');
   });
 
   it('recognizes invalid amount', () => {
     wrapper.find('.amount input').simulate('change', { target: { value: '120 INVALID' } });
-    expect(wrapper.find('.amount').text()).to.contain('Invalid');
+    expect(wrapper.find('Input.amount').text()).to.contain('Invalid');
   });
 
   it('recognizes zero amount', () => {
     wrapper.find('.amount input').simulate('change', { target: { value: '0' } });
-    expect(wrapper.find('.amount').text()).to.contain('Zero not allowed');
+    expect(wrapper.find('Input.amount').text()).to.contain('Zero not allowed');
   });
 
   it('recognizes too high amount', () => {
     wrapper.find('.amount input').simulate('change', { target: { value: '12000' } });
-    expect(wrapper.find('.amount').text()).to.contain('Insufficient funds');
+    expect(wrapper.find('Input.amount').text()).to.contain('Insufficient funds');
   });
 
   it('recognizes empty amount', () => {
     wrapper.find('.amount input').simulate('change', { target: { value: '12000' } });
     wrapper.find('.amount input').simulate('change', { target: { value: '' } });
-    expect(wrapper.find('.amount').text()).to.contain('Required');
+    expect(wrapper.find('Input.amount').text()).to.contain('Required');
   });
 
   it('accepts valid recipient', () => {
     wrapper.find('.recipient input').simulate('change', { target: { value: '11004588490103196952L' } });
-    expect(wrapper.find('.recipient').text()).to.not.contain('Invalid');
+    expect(wrapper.find('Input.recipient').text()).to.not.contain('Invalid');
   });
 
   it('recognizes invalid recipient', () => {
     wrapper.find('.recipient input').simulate('change', { target: { value: '11004588490103196952' } });
-    expect(wrapper.find('.recipient').text()).to.contain('Invalid');
+    expect(wrapper.find('Input.recipient').text()).to.contain('Invalid');
   });
 
   it('allows to set maximum amount', () => {
-    wrapper.find('.transaction-amount').simulate('click');
-    wrapper.find('.send-maximum-amount').simulate('click');
-    expect(wrapper.find('.amount input').props().value).to.equal('999.9');
+    wrapper.find('IconMenu.transaction-amount').simulate('click');
+    wrapper.find('MenuItem.send-maximum-amount').simulate('click');
+    expect(wrapper.find('Input.amount input').props().value).to.equal('999.9');
   });
 
   it('allows to send a transaction', () => {
