@@ -15,7 +15,7 @@ const TransactionType = (props) => {
       type = t('Delegate Registration');
       break;
     case 3:
-      type = t('Vote');
+      type = t('Vote', { context: 'noun' });
       break;
     case 4:
       type = t('Multisignature Creation');
