@@ -8,12 +8,16 @@ import i18n from '../../i18n';
 import { VoteAutocompleteRaw as VoteAutocomplete } from './voteAutocomplete';
 
 const votes = {
-  username1: { publicKey: 'sample_key', confirmed: true, unconfirmed: false },
-  username2: { publicKey: 'sample_key', confirmed: false, unconfirmed: true },
+  username1: { publicKey: 'sample_key_1', confirmed: true, unconfirmed: false },
+  username2: { publicKey: 'sample_key_2', confirmed: false, unconfirmed: true },
 };
 const delegates = [
   { username: 'username1', publicKey: '123HG3452245L' },
   { username: 'username2', publicKey: '123HG3522345L' },
+];
+const unvotedDelegate = [
+  { username: 'username3', publicKey: '123HG3522445L' },
+  { username: 'username4', publicKey: '123HG3522545L' },
 ];
 const props = {
   activePeer: {},
@@ -23,6 +27,12 @@ const props = {
   t: key => key,
 };
 let wrapper;
+const keyCodes = {
+  arrowDown: 40,
+  arrowUp: 38,
+  enter: 13,
+  escape: 27,
+};
 
 const store = configureMockStore([])({
   peers: {},
@@ -36,8 +46,12 @@ const store = configureMockStore([])({
 describe('VoteAutocomplete', () => {
   let voteAutocompleteApiMock;
   let unvoteAutocompleteApiMock;
+  let clock;
 
   beforeEach(() => {
+    clock = sinon.useFakeTimers({
+      toFake: ['setTimeout', 'clearTimeout', 'Date'],
+    });
     sinon.spy(VoteAutocomplete.prototype, 'keyPress');
     sinon.spy(VoteAutocomplete.prototype, 'handleArrowDown');
     sinon.spy(VoteAutocomplete.prototype, 'handleArrowUp');
@@ -48,6 +62,7 @@ describe('VoteAutocomplete', () => {
   });
 
   afterEach(() => {
+    clock.restore();
     voteAutocompleteApiMock.restore();
     unvoteAutocompleteApiMock.restore();
     VoteAutocomplete.prototype.keyPress.restore();
@@ -55,34 +70,27 @@ describe('VoteAutocomplete', () => {
     VoteAutocomplete.prototype.handleArrowUp.restore();
   });
 
-  it('should suggestionStatus(false, className) change value of className in state', () => {
-    const clock = sinon.useFakeTimers();
-    wrapper.instance().suggestionStatus(false, 'className');
-    clock.tick(200);
-    expect(wrapper.state('className').match(/hidden/g)).to.have.lengthOf(1);
+  it('should suggest with full username if finds a non-voted delegate with a username starting with given string', () => {
+    const voteAutocompleteApiStub = sinon.stub(delegateApi, 'voteAutocomplete');
+    voteAutocompleteApiStub.returnsPromise().resolves([unvotedDelegate[0]]);
+    wrapper.find('.votedListSearch.vote-auto-complete input').simulate('change', { target: { value: 'user' } });
+
+    clock.tick(300);
+    expect(wrapper.find('Card .vote-auto-complete-list ul').html().indexOf(unvotedDelegate[0].username)).to.be.greaterThan(-1);
+    voteAutocompleteApiStub.restore();
   });
 
-  it('should suggestionStatus(true, className) clear value of className in state', () => {
-    const clock = sinon.useFakeTimers();
-    wrapper.instance().suggestionStatus(true, 'className');
-    clock.tick(200);
-    expect(wrapper.state('className')).to.be.equal('');
+  it('should suggest with full username if dows not find a non-voted delegate with a username starting with given string', () => {
+    const voteAutocompleteApiStub = sinon.stub(delegateApi, 'voteAutocomplete');
+    voteAutocompleteApiStub.returnsPromise().rejects([]);
+    wrapper.find('.votedListSearch.vote-auto-complete input').simulate('change', { target: { value: 'user' } });
+
+    clock.tick(400);
+    expect(wrapper.find('Card .vote-auto-complete-list ul')).to.be.blank();
+    voteAutocompleteApiStub.restore();
   });
 
-  it('should search hide suggestion boxes when value is equal to ""', () => {
-    const clock = sinon.useFakeTimers();
-    sinon.spy(VoteAutocomplete.prototype, 'setState');
-    wrapper.instance().search('votedListSearch', '');
-    clock.tick(250);
-    expect(VoteAutocomplete.prototype.setState).to.have.property('callCount', 2);
-    expect(wrapper.state('votedResult')).to.have.lengthOf(0);
-    expect(wrapper.state('unvotedResult')).to.have.lengthOf(0);
-    expect(wrapper.state('votedSuggestionClass').match(/hidden/g)).to.have.lengthOf(1);
-    expect(wrapper.state('unvotedSuggestionClass').match(/hidden/g)).to.have.lengthOf(1);
-    VoteAutocomplete.prototype.setState.restore();
-  });
   it('search should call "voteAutocomplete" when name is equal to "votedListSearch" when search term exists', () => {
-    const clock = sinon.useFakeTimers();
     const existingSearchTerm = 'username2';
     const delegateApiMock = sinon.mock(delegateApi).expects('voteAutocomplete');
 
@@ -97,7 +105,6 @@ describe('VoteAutocomplete', () => {
   });
 
   it('search should call "voteAutocomplete" when name is equal to "votedListSearch" when search term does not exist', () => {
-    const clock = sinon.useFakeTimers();
     const nonExistingSearchTerm = 'doesntexist';
     const delegateApiMock = sinon.mock(delegateApi).expects('voteAutocomplete');
 
@@ -113,7 +120,6 @@ describe('VoteAutocomplete', () => {
   });
 
   it('search should call "unvoteAutocomplete" when name is equal to "unvotedListSearch" when search term exists', () => {
-    const clock = sinon.useFakeTimers();
     const existingSearchTerm = 'username1';
     const delegateApiMock = sinon.mock(delegateApi).expects('unvoteAutocomplete');
 
@@ -128,7 +134,6 @@ describe('VoteAutocomplete', () => {
   });
 
   it('search should call "unvoteAutocomplete" when name is equal to "unvotedListSearch" when search term does not exists', () => {
-    const clock = sinon.useFakeTimers();
     const nonExistingSearchTerm = 'username2';
     const delegateApiMock = sinon.mock(delegateApi).expects('unvoteAutocomplete');
 
@@ -141,81 +146,119 @@ describe('VoteAutocomplete', () => {
 
     delegateApiMock.restore();
   });
-  it('should "votedSearchKeydown" call "keyPress"', () => {
-    wrapper.instance().votedSearchKeyDown({});
-    expect(VoteAutocomplete.prototype.keyPress).to.have.property('callCount', 1);
-  });
 
-  it('should "unvotedSearchKeydown" call "keyPress"', () => {
-    wrapper.instance().unvotedSearchKeyDown({});
-    expect(VoteAutocomplete.prototype.keyPress).to.have.property('callCount', 1);
-  });
+  it('should let you choose one of the options by arrow down', () => {
+    const voteAutocompleteApiStub = sinon.stub(delegateApi, 'voteAutocomplete');
+    voteAutocompleteApiStub.returnsPromise().resolves(unvotedDelegate);
+    // write a username
+    wrapper.find('.votedListSearch.vote-auto-complete input').simulate('change', { target: { value: 'user' } });
+    clock.tick(400);
 
-  it('should keyPress call "handleArrowDown" when event.keyCode is equal to 40', () => {
-    const list = [
-      { address: 'address 0' },
-      { address: 'address 1', hovered: true },
-      { address: 'address 2' },
-    ];
-    wrapper.setState({ votedResult: list });
-    wrapper.instance().keyPress({ keyCode: 40 }, 'votedSuggestionClass', 'votedResult');
-    expect(VoteAutocomplete.prototype.handleArrowDown).to.have.property('callCount', 1);
-  });
-  it('should handleArrowDown select first item in votedResult when no one is selected', () => {
-    const list = [
-      { address: 'address 0' },
-      { address: 'address 1' },
-      { address: 'address 2' },
-    ];
-    wrapper.setState({ votedResult: list });
-    wrapper.instance().handleArrowDown(wrapper.state('votedResult'), 'votedResult');
-    expect(wrapper.state('votedResult')[0].hovered).to.have.be.equal(true);
-  });
-
-  it('should handleArrowDown select first item in unvotedResult when no one is selected', () => {
-    const list = [
-      { address: 'address 0' },
-      { address: 'address 1' },
-      { address: 'address 2' },
-    ];
-    wrapper.setState({ unvotedResult: list });
-    wrapper.instance().handleArrowDown(wrapper.state('unvotedResult'), 'unvotedResult');
-    expect(wrapper.state('unvotedResult')[0].hovered).to.have.be.equal(true);
-  });
-  it('should keyPress call "handleArrowUp" when event.keyCode is equal to 38', () => {
-    const list = [
-      { address: 'address 0' },
-      { address: 'address 1', hovered: true },
-    ];
-    wrapper.setState({ votedResult: list });
-    wrapper.instance().keyPress({ keyCode: 38 }, 'votedSuggestionClass', 'votedResult');
-    expect(VoteAutocomplete.prototype.handleArrowUp).to.have.property('callCount', 1);
-  });
-
-  it('should keyPress hide suggestions when event.keyCode is equal to 27', () => {
-    const returnValue = wrapper.instance()
-      .keyPress({ keyCode: 27 }, 'votedSuggestionClass', 'votedResult');
-    expect(wrapper.state('votedSuggestionClass').match(/hidden/g)).to.have.lengthOf(1);
-    expect(returnValue).to.be.equal(false);
-  });
-
-  it(`should keyPress call "addToVoted" when event.keyCode is equal to 13 and 
-    list name is equal to votedResult`, () => {
-      const list = [{ address: 'address 1', hovered: true }];
-      wrapper.setState({ votedResult: list });
-      const returnValue = wrapper.instance()
-        .keyPress({ keyCode: 13 }, 'votedSuggestionClass', 'votedResult');
-      expect(props.voteToggled).to.have.property('callCount', 1);
-      expect(returnValue).to.be.equal(false);
+    // select it with arrow down
+    wrapper.find('.votedListSearch.vote-auto-complete input').simulate('keyDown', { keyCode: keyCodes.arrowDown });
+    clock.tick(200);
+    wrapper.find('.votedListSearch.vote-auto-complete input').simulate('keyDown', { keyCode: keyCodes.enter });
+    voteAutocompleteApiStub.restore();
+    expect(props.voteToggled).to.have.been.calledWith({
+      publicKey: unvotedDelegate[0].publicKey,
+      username: unvotedDelegate[0].username,
     });
+  });
 
-  it(`should keyPress call "voteToggled" when event.keyCode is equal to 13 and 
-    list name is equal to unvotedResult`, () => {
-      const list = [{ address: 'address 1', hovered: true }];
-      wrapper.setState({ unvotedResult: list });
-      const returnValue = wrapper.instance()
-        .keyPress({ keyCode: 13 }, 'unvotedSuggestionClass', 'unvotedResult');
-      expect(props.voteToggled).to.have.property('callCount', 2);
-      expect(returnValue).to.be.equal(false);
+  it('should let you navigate and choose one of the options in voteList using arrow up/down', () => {
+    const voteAutocompleteApiStub = sinon.stub(delegateApi, 'voteAutocomplete');
+    voteAutocompleteApiStub.returnsPromise().resolves(unvotedDelegate);
+    // write a username
+    wrapper.find('.votedListSearch.vote-auto-complete input').simulate('change', { target: { value: 'user' } });
+    clock.tick(400);
+
+    // Arrow down
+    wrapper.find('.votedListSearch.vote-auto-complete input').simulate('keyDown', { keyCode: keyCodes.arrowDown });
+    // Arrow down
+    wrapper.find('.votedListSearch.vote-auto-complete input').simulate('keyDown', { keyCode: keyCodes.arrowDown });
+    // Arrow up
+    wrapper.find('.votedListSearch.vote-auto-complete input').simulate('keyDown', { keyCode: keyCodes.arrowUp });
+    // Hit enter
+    wrapper.find('.votedListSearch.vote-auto-complete input').simulate('keyDown', { keyCode: keyCodes.enter });
+    voteAutocompleteApiStub.restore();
+    expect(props.voteToggled).to.have.been.calledWith({
+      publicKey: unvotedDelegate[0].publicKey,
+      username: unvotedDelegate[0].username,
     });
+  });
+
+  it('should let you navigate and then escape the suggestion list', () => {
+    const voteAutocompleteApiStub = sinon.stub(delegateApi, 'voteAutocomplete');
+    voteAutocompleteApiStub.returnsPromise().resolves(unvotedDelegate);
+    // write a username
+    wrapper.find('.votedListSearch.vote-auto-complete input').simulate('change', { target: { value: 'user' } });
+    clock.tick(400);
+
+    // Arrow down
+    wrapper.find('.votedListSearch.vote-auto-complete input').simulate('keyDown', { keyCode: keyCodes.arrowDown });
+    // Arrow down
+    wrapper.find('.votedListSearch.vote-auto-complete input').simulate('keyDown', { keyCode: keyCodes.arrowDown });
+    // Arrow up
+    wrapper.find('.votedListSearch.vote-auto-complete input').simulate('keyDown', { keyCode: keyCodes.arrowUp });
+    // Hit enter
+    wrapper.find('.votedListSearch.vote-auto-complete input').simulate('keyDown', { keyCode: keyCodes.escape });
+    voteAutocompleteApiStub.restore();
+    clock.tick(400);
+    expect(wrapper.find('Card .vote-auto-complete-list').at(0).prop('className')).to.include('hidden');
+  });
+
+  it('should remove suggestion list if you clean the input', () => {
+    const voteAutocompleteApiStub = sinon.stub(delegateApi, 'voteAutocomplete');
+    voteAutocompleteApiStub.returnsPromise().resolves(unvotedDelegate);
+    // write a username
+    wrapper.find('.votedListSearch.vote-auto-complete input').simulate('change', { target: { value: 'user' } });
+    clock.tick(400);
+    wrapper.update();
+    wrapper.find('.votedListSearch.vote-auto-complete input').simulate('change', { target: { value: '' } });
+    voteAutocompleteApiStub.restore();
+    clock.tick(400);
+    wrapper.update();
+    expect(wrapper.find('Card .vote-auto-complete-list ul')).to.be.blank();
+  });
+
+  it('should hide suggestion list if you blur the input', () => {
+    const voteAutocompleteApiStub = sinon.stub(delegateApi, 'voteAutocomplete');
+    voteAutocompleteApiStub.returnsPromise().resolves(unvotedDelegate);
+    // write a username
+    wrapper.find('.votedListSearch.vote-auto-complete input').simulate('change', { target: { value: 'user' } });
+    clock.tick(400);
+    wrapper.find('.votedListSearch.vote-auto-complete input').simulate('blur', {});
+    voteAutocompleteApiStub.restore();
+    clock.tick(200);
+    wrapper.update();
+    expect(wrapper.find('Card .vote-auto-complete-list').at(0).prop('className')).to.include('hidden');
+  });
+
+  it('should suggest with full username to unvote if finds a voted delegate with a username starting with given string', () => {
+    const unvoteAutocompleteApiStub = sinon.stub(delegateApi, 'unvoteAutocomplete');
+    unvoteAutocompleteApiStub.returnsPromise().resolves([delegates[1]]);
+    wrapper.find('.unvotedListSearch input').simulate('change', { target: { value: 'user' } });
+
+    clock.tick(300);
+    expect(wrapper.find('Card .unvote-auto-complete-list ul').html()).to.include(delegates[1].username);
+    unvoteAutocompleteApiStub.restore();
+  });
+
+  it('should let you navigate and choose one of the options in unvoteList using arrow up/down', () => {
+    const unvoteAutocompleteApiStub = sinon.stub(delegateApi, 'unvoteAutocomplete');
+    unvoteAutocompleteApiStub.returnsPromise().resolves([delegates[1]]);
+    // write a username
+    wrapper.find('.unvotedListSearch input').simulate('change', { target: { value: 'user' } });
+    clock.tick(400);
+
+    // Arrow down
+    wrapper.find('.unvotedListSearch input').simulate('keyDown', { keyCode: keyCodes.arrowDown });
+    // Hit enter
+    wrapper.find('.unvotedListSearch input').simulate('keyDown', { keyCode: keyCodes.enter });
+    unvoteAutocompleteApiStub.restore();
+    expect(props.voteToggled).to.have.been.calledWith({
+      publicKey: delegates[1].publicKey,
+      username: delegates[1].username,
+    });
+  });
 });
