@@ -1,10 +1,9 @@
 import { expect } from 'chai';
 import { spy, stub } from 'sinon';
-
-import { SYNC_ACTIVE_INTERVAL, SYNC_INACTIVE_INTERVAL } from '../../constants/api';
 import { accountUpdated } from '../../actions/account';
 import { activePeerUpdate } from '../../actions/peers';
 import * as votingActions from '../../actions/voting';
+import * as forgingActions from '../../actions/forging';
 import * as accountApi from '../../utils/api/account';
 import actionTypes from '../../constants/actions';
 import * as delegateApi from '../../utils/api/delegate';
@@ -19,6 +18,8 @@ describe('Account middleware', () => {
   let stubTransactions;
   const passphrase = 'right cat soul renew under climb middle maid powder churn cram coconut';
 
+  const transactions = { transactions: [{ senderId: 'sample_address', receiverId: 'some_address' }] };
+
   const transactionsUpdatedAction = {
     type: actionTypes.transactionsUpdated,
     data: {
@@ -29,17 +30,19 @@ describe('Account middleware', () => {
     },
   };
 
-  const activeBeatAction = {
-    type: actionTypes.metronomeBeat,
+  const newBlockCreated = {
+    type: actionTypes.newBlockCreated,
     data: {
-      interval: SYNC_ACTIVE_INTERVAL,
+      windowIsFocused: true,
+      block: transactions,
     },
   };
 
-  const inactiveBeatAction = {
-    type: actionTypes.metronomeBeat,
+  const inactiveNewBlockCreated = {
+    type: actionTypes.newBlockCreated,
     data: {
-      interval: SYNC_INACTIVE_INTERVAL,
+      windowIsFocused: false,
+      block: transactions,
     },
   };
 
@@ -74,85 +77,88 @@ describe('Account middleware', () => {
   });
 
   it('should pass the action to next middleware', () => {
-    const expectedAction = {
-      type: 'TEST_ACTION',
-    };
-
-    middleware(store)(next)(expectedAction);
-    expect(next).to.have.been.calledWith(expectedAction);
+    middleware(store)(next)(newBlockCreated);
+    expect(next).to.have.been.calledWith(newBlockCreated);
   });
 
-  it(`should call account API methods on ${actionTypes.metronomeBeat} action when online`, () => {
+  it(`should call account API methods on ${actionTypes.newBlockCreated} action when online`, () => {
+    // does this matter?
     stubGetAccount.resolves({ balance: 0 });
 
-    middleware(store)(next)(activeBeatAction);
+    middleware(store)(next)(newBlockCreated);
 
     expect(stubGetAccount).to.have.been.calledWith();
     expect(store.dispatch).to.have.been.calledWith(activePeerUpdate({ online: true }));
   });
 
-  it(`should call account API methods on ${actionTypes.metronomeBeat} action when offline`, () => {
+  it(`should call account API methods on ${actionTypes.newBlockCreated} action when offline`, () => {
     const errorCode = 'EUNAVAILABLE';
     stubGetAccount.rejects({ error: { code: errorCode } });
 
-    middleware(store)(next)(activeBeatAction);
+    middleware(store)(next)(newBlockCreated);
 
     expect(store.dispatch).to.have.been.calledWith(activePeerUpdate(
       { online: false, code: errorCode }));
   });
 
-  it(`should call transactions API methods on ${actionTypes.metronomeBeat} action if account.balance changes`, () => {
+  it(`should call transactions API methods on ${actionTypes.newBlockCreated} action if account.balance changes`, () => {
     stubGetAccount.resolves({ balance: 10e8 });
 
-    middleware(store)(next)(activeBeatAction);
+    middleware(store)(next)(newBlockCreated);
 
     expect(stubGetAccount).to.have.been.calledWith();
-    // TODO why next expect doesn't work despite it being called according to test coverage?
-    // expect(stubTransactions).to.have.been.calledWith();
+    expect(stubTransactions).to.have.been.calledWith();
   });
 
-  it(`should call transactions API methods on ${actionTypes.metronomeBeat} action if account.balance changes and action.data.interval is SYNC_INACTIVE_INTERVAL`, () => {
+  it(`should call transactions API methods on ${actionTypes.newBlockCreated} action if account.balance changes and the window is in blur`, () => {
     stubGetAccount.resolves({ balance: 10e8 });
 
-    middleware(store)(next)(inactiveBeatAction);
+    middleware(store)(next)(inactiveNewBlockCreated);
 
     expect(stubGetAccount).to.have.been.calledWith();
-    // TODO why next expect doesn't work despite it being called according to test coverage?
-    // expect(stubTransactions).to.have.been.calledWith();
+    expect(stubTransactions).to.have.been.calledWith();
   });
 
-  it(`should call transactions API methods on ${actionTypes.metronomeBeat} action if action.data.interval is SYNC_ACTIVE_INTERVAL and there are recent transactions`, () => {
+  it(`should call transactions API methods on ${actionTypes.newBlockCreated} action if account.balance changes the user has no transactions yet`, () => {
+    stubGetAccount.resolves({ balance: 10e8 });
+
+    state.transactions.count = 0;
+    middleware(store)(next)(newBlockCreated);
+
+    expect(stubGetAccount).to.have.been.calledWith();
+    // eslint-disable-next-line no-unused-expressions
+    expect(stubTransactions).to.have.been.calledOnce;
+  });
+
+  it(`should call transactions API methods on ${actionTypes.newBlockCreated} action if the window is in focus and there are recent transactions`, () => {
     stubGetAccount.resolves({ balance: 0 });
 
-    middleware(store)(next)(activeBeatAction);
+    middleware(store)(next)(newBlockCreated);
 
     expect(stubGetAccount).to.have.been.calledWith();
-    // TODO why next expect doesn't work despite it being called according to test coverage?
-    // expect(stubTransactions).to.have.been.calledWith();
+    expect(stubTransactions).to.have.been.calledWith();
   });
 
-  it(`should fetch delegate info on ${actionTypes.metronomeBeat} action if account.balance changes and account.isDelegate`, () => {
+  it(`should fetch delegate info on ${actionTypes.newBlockCreated} action if account.balance changes and account.isDelegate`, () => {
     const delegateApiMock = stub(delegateApi, 'getDelegate').returnsPromise().resolves({ success: true, delegate: {} });
     stubGetAccount.resolves({ balance: 10e8 });
     state.account.isDelegate = true;
     store.getState = () => (state);
 
-    middleware(store)(next)(activeBeatAction);
+    middleware(store)(next)(newBlockCreated);
     expect(store.dispatch).to.have.been.calledWith();
 
     delegateApiMock.restore();
   });
 
-  it(`should call fetchAndUpdateForgedBlocks(...) on ${actionTypes.metronomeBeat} action if account.balance changes and account.isDelegate`, () => {
+  it(`should call fetchAndUpdateForgedBlocks(...) on ${actionTypes.newBlockCreated} action if account.balance changes and account.isDelegate`, () => {
     state.account.isDelegate = true;
     store.getState = () => (state);
     stubGetAccount.resolves({ balance: 10e8 });
-    // const fetchAndUpdateForgedBlocksSpy = spy(forgingActions, 'fetchAndUpdateForgedBlocks');
+    const fetchAndUpdateForgedBlocksSpy = spy(forgingActions, 'fetchAndUpdateForgedBlocks');
 
-    middleware(store)(next)({ type: actionTypes.metronomeBeat });
-
-    // TODO why next expect doesn't work despite it being called according to test coverage?
-    // expect(fetchAndUpdateForgedBlocksSpy).to.have.been.calledWith();
+    middleware(store)(next)(newBlockCreated);
+    expect(fetchAndUpdateForgedBlocksSpy).to.have.been.calledWith();
   });
 
   it(`should fetch delegate info on ${actionTypes.transactionsUpdated} action if action.data.confirmed contains delegateRegistration transactions`, () => {
