@@ -1,5 +1,4 @@
 import thunk from 'redux-thunk';
-import metronomeMiddleware from './metronome';
 import accountMiddleware from './account';
 import loginMiddleware from './login';
 import transactionsMiddleware from './transactions';
@@ -8,12 +7,13 @@ import offlineMiddleware from './offline';
 import notificationMiddleware from './notification';
 import votingMiddleware from './voting';
 import savedAccountsMiddleware from './savedAccounts';
+import socketMiddleware from './socket';
 
 export default [
   thunk,
   transactionsMiddleware,
   loginMiddleware,
-  metronomeMiddleware,
+  socketMiddleware,
   accountMiddleware,
   loadingBarMiddleware,
   offlineMiddleware,
