@@ -88,6 +88,14 @@ export const generatePassphrase = ({ seed }) => (new mnemonic(new Buffer(seed.jo
    * @returns {bool} isValidPassphrase
    */
 export const isValidPassphrase = (passphrase) => {
-  const normalizedValue = passphrase.replace(/ +/g, ' ').trim().toLowerCase();
-  return normalizedValue.split(' ').length >= 12 && mnemonic.isValid(normalizedValue);
+  const normalizedValue = passphrase.replace(/ +/g, ' ').trim();
+  let isValid;
+  try {
+    isValid = normalizedValue.split(' ').length >= 12 && mnemonic.isValid(normalizedValue);
+  } catch (e) {
+    // If the mnemonic check throws an error, we assume that the
+    // passphrase being entered isn't valid
+    isValid = false;
+  }
+  return isValid;
 };
